chore(api): remove legacy router helpers
