+++ openrag/api/routers/admin/cluster.py
@@ -1,8 +1,8 @@
+from api.dependencies.auth import require_admin
 from di.workers import list_ray_actors as list_ray_actor_states
 from di.workers import restart_ray_actor
 from fastapi import APIRouter, Depends, HTTPException, status
 from fastapi.responses import JSONResponse
-from routers.utils import require_admin
 from utils.logger import get_logger
 
 logger = get_logger()
