+++ openrag/api/routers/admin/indexing.py
@@ -15,6 +15,15 @@ import json
 from pathlib import Path
 from typing import Any
 
+from api.dependencies.auth import (
+    check_user_file_quota,
+    current_user_partitions,
+    ensure_partition_role,
+    require_partition_editor,
+    require_task_owner,
+)
+from api.dependencies.files import validate_file_format, validate_file_id, validate_metadata
+from api.routers.admin.task_logs import collect_task_logs
 from components.indexer.utils.files import sanitize_filename, save_file_to_disk
 from config import load_config
 from di.providers import get_auth_service, get_indexing_service, get_partition_service
@@ -29,17 +38,6 @@ from fastapi import (
     status,
 )
 from fastapi.responses import JSONResponse
-from routers.task_logs import collect_task_logs
-from routers.utils import (
-    check_user_file_quota,
-    current_user_partitions,
-    ensure_partition_role,
-    require_partition_editor,
-    require_task_owner,
-    validate_file_format,
-    validate_file_id,
-    validate_metadata,
-)
 from utils.logger import get_logger
 
 logger = get_logger()
