+++ openrag/api/routers/user/search.py
@@ -11,14 +11,14 @@ shaping (domain ``Chunk`` → ``{link, metadata, content}``).
 
 from typing import Annotated
 
-from di.providers import get_retrieval_service, get_workspace_service
-from fastapi import APIRouter, Depends, HTTPException, Query, Request, status
-from fastapi.responses import JSONResponse
-from routers.utils import (
+from api.dependencies.auth import (
     current_user_or_admin_partitions_list,
     require_partition_viewer,
     require_partitions_viewer,
 )
+from di.providers import get_retrieval_service, get_workspace_service
+from fastapi import APIRouter, Depends, HTTPException, Query, Request, status
+from fastapi.responses import JSONResponse
 from utils.logger import get_logger
 
 logger = get_logger()
